requirements.txt
streamlit>=1.20.0
ollama>=0.1.0
Pillow>=9.0.0
